+++ test/jet/main_test.clj
@@ -153,3 +153,7 @@
 (deftest specter-test
   (is (= "[{:a {:aa 2}, :b {:ba 0, :bb 3}} [3 3 18 6 12] [2 1 3 6 10 4 8] {:a [1 2 3]} {:a {:b {}}} {} [0 2 2 4 4 5 6 7] [0 1 :a :b :c :d :e 4 5 6 7 8 9] [[1 :a :b] (1 2 :a :b) [:c :a :b]] [2 1 2 6 7 4 1 2] [10] [0 1 2 3 10 5 8 7 6 9 4 11 12 13 14 15] [[1 2 3 4 5 6 :c :d] [7 0 -1] [8 8 :c :d] []] [{:a 1, :b 3} {:a -8, :b -10} {:a 14, :b 10} {:a 3}] {:a 11, :b 3} [{:a 2, :c [2 3], :d 4} {:a 4, :c [1 11 0]} {:a -1, :c [1 1 1], :d 2}]]\n"
          (jet "{:a {:a 1}}" "--no-pretty" "-f" "test-resources/specter-test.clj"))))
+
+(deftest base64-test
+  (is (= "\"Zm9v\"\n" (jet "{:a \"foo\"}" "-t" ":a base64/encode")))
+  (is (= "\"foo\"\n" (jet "{:a \"foo\"}" "-t" ":a base64/encode base64/decode"))))