+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.4.22 (2023.01.27)
+
+- [#123](https://github.com/borkdude/jet/issues/123): Add `base64/encode` and `base64/decode`
+- Deprecate interactive mode
+- Deprecate `--query` in favor of `--thread-last`, `--thread-first` or `--func`
+
 ## 0.3.21 (2022.10.28)
 
 - [#74](https://github.com/borkdude/jet/issues/74): Add yaml support ([@qdzo](https://github.com/qdzo))